+++ packages/core/src/environment.ts
@@ -38,12 +38,13 @@ export function supportedEnvironments(): Environment[] {
   return environments
 }
 
-export function resolveEnvironment(environment_id: string): ResolvedEnvironment {
+export function resolveEnvironment(environment_id: string, customProvider?: string): ResolvedEnvironment {
   const protocolConfig = require('../protocol-config.json') as ProtocolConfig
   const environment = protocolConfig.environments[environment_id]
   const network = protocolConfig.networks[environment?.network_id]
   if (environment && network) {
     network.id = environment?.network_id
+    network.default_provider = customProvider ?? network?.default_provider
     return {
       id: environment_id,
       network,