+++ packages/core/src/main.ts
@@ -3,7 +3,7 @@ import { mkdir } from 'fs/promises'
 
 import { default as LibP2P, type Connection } from 'libp2p'
 import { LevelDatastore } from 'datastore-level'
-import { type AddressSorter, expandVars, HoprDB, PublicKey, debug } from '@hoprnet/hopr-utils'
+import { type AddressSorter, HoprDB, PublicKey, debug } from '@hoprnet/hopr-utils'
 import HoprCoreEthereum from '@hoprnet/hopr-core-ethereum'
 import Mplex from 'libp2p-mplex'
 import KadDHT from 'libp2p-kad-dht'
@@ -190,8 +190,7 @@ export async function createHoprNode(
     throw err
   }
 
-  const provider = expandVars(options.environment.network.default_provider, process.env)
-  log(`using provider URL: ${provider}`)
+  log(`using provider URL: ${options.environment.network.default_provider}`)
   const chain = new HoprCoreEthereum(
     db,
     PublicKey.fromPeerId(peerId),
@@ -201,7 +200,7 @@ export async function createHoprNode(
       environment: options.environment.id,
       gasPrice: options.environment.network.gas_price,
       network: options.environment.network.id,
-      provider
+      provider: options.environment.network.default_provider
     },
     automaticChainCreation
   )